requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=0.24.0
cryptography>=3.4.0
pycryptodome>=3.10.0
matplotlib>=3.4.0